requirements.txt
discord.py[voice]
Flask
Flask-Limiter
python-dotenv
pymysql
colorlog
